Move canvasSize from image.go to size_windows.go and size_unix.go in preparation for changes to make it build and work on windows.
