+++ size_unix.go
@@ -1,0 +1,24 @@
+// +build !windows
+package main
+
+import (
+	"os"
+	"syscall"
+	"unsafe"
+)
+
+// canvasSize returns the terminal columns, rows, and cursor aspect ratio
+func canvasSize() (int, int, float64) {
+	var size [4]uint16
+	if _, _, err := syscall.Syscall6(syscall.SYS_IOCTL, uintptr(os.Stdout.Fd()), uintptr(syscall.TIOCGWINSZ), uintptr(unsafe.Pointer(&size)), 0, 0, 0); err != 0 {
+		panic(err)
+	}
+	rows, cols, width, height := size[0], size[1], size[2], size[3]
+
+	var whratio = defaultRatio
+	if width > 0 && height > 0 {
+		whratio = float64(height/rows) / float64(width/cols)
+	}
+
+	return int(cols), int(rows), whratio
+}
